works

+++ src/random_population.rs
@@ -1,6 +1,6 @@
 use rand::Rng;
 
-fn create_individual(n_items: u32) -> Vec<bool> {
+fn create_individual(n_items: u16) -> Vec<bool> {
     let mut ind = Vec::with_capacity(n_items as usize);
     for _i in 0..n_items {
         let nmb = rand::thread_rng().gen_range(0,2);
@@ -17,7 +17,7 @@ fn create_individual(n_items: u32) -> Vec<bool> {
  * n_items - number of items from which the subset tobe packed into the knapsackis selected. numberof items generated in task 1
  * size - population size
  * */
-pub fn init_population(n_items: u32, size: u32) -> Vec<Vec<bool>> {
+pub fn init_population(n_items: u16, size: u16) -> Vec<Vec<bool>> {
     let mut population = Vec::<Vec<bool>>::with_capacity(size as usize);
     for _i in 0..size {
         population.push(create_individual(n_items));
